15439 POSProduct networking tests in SwiftTesting
