+++ Networking/NetworkingTests/Remote/POSProductsNetworkingTests.swift
@@ -1,0 +1,133 @@
+import Testing
+@testable import Networking
+
+struct POSProductsNetworkingTests {
+
+    let network = MockNetwork()
+    let sampleSiteID: Int64 = 1234
+
+    @Test func loadProductsForPointOfSale_loads_simple_products() async throws {
+        // Given
+        let remote = ProductsRemote(network: network)
+        let expectedProductsFromResponse = 6
+
+        // When
+        network.simulateResponse(requestUrlSuffix: "products", filename: "products-load-all-type-simple")
+
+        let pagedProducts = try await remote.loadProductsForPointOfSale(for: sampleSiteID)
+
+        // Then
+        let products = pagedProducts.items
+        #expect(products.count == expectedProductsFromResponse)
+
+        for product in products {
+            #expect(product.productType == .simple)
+        }
+    }
+
+    @Test func loadProductsForPointOfSale_relays_networking_error() async throws {
+        // Given
+        let remote = ProductsRemote(network: network)
+
+        // When/Then
+        await #expect(throws: NetworkError.notFound()) {
+            try await remote.loadProductsForPointOfSale(for: sampleSiteID)
+        }
+    }
+
+    @Test func loadProductsForPointOfSale_when_page_has_products_then_loads_expected_products() async throws {
+        // Given
+        let remote = ProductsRemote(network: network)
+        let initialPageNumber = 1
+        let expectedProductsFromResponse = 6
+
+        // When
+        network.simulateResponse(requestUrlSuffix: "products", filename: "products-load-all-type-simple")
+
+        let pagedProducts = try await remote.loadProductsForPointOfSale(for: sampleSiteID, pageNumber: initialPageNumber)
+
+        // Then
+        let products = pagedProducts.items
+        #expect(products.count == expectedProductsFromResponse)
+
+        for product in products {
+            #expect(product.productType == .simple)
+        }
+    }
+
+    @Test func loadProductsForPointOfSale_sets_correct_parameters() async throws {
+        // Given
+        let remote = ProductsRemote(network: network)
+
+        // When
+        _ = try? await remote.loadProductsForPointOfSale(for: sampleSiteID, productTypes: [.simple, .variable], pageNumber: 1)
+
+        // Then
+        let queryParametersDictionary = try #require(network.queryParametersDictionary as? [String: String])
+        #expect(queryParametersDictionary["downloadable"] == "false")
+        #expect(queryParametersDictionary["include_types"] == "simple,variable")
+    }
+
+    @Test(arguments: 1...4) func loadProductsForPointOfSale_returns_hasMorePages_based_on_header_with_case_insensitive_name(pageNumber: Int) async throws {
+        // Given a response with 5 pages
+        let remote = ProductsRemote(network: network)
+        network.responseHeaders = ["X-WP-TotalPages": "5"]
+        network.simulateResponse(requestUrlSuffix: "products", filename: "empty-data-array")
+
+        // When loading page 1 to 4
+        let pagedProducts = try await remote.loadProductsForPointOfSale(
+            for: sampleSiteID,
+            pageNumber: pageNumber)
+
+        // Then there are more pages
+        #expect(pagedProducts.hasMorePages == true)
+    }
+
+    @Test(arguments: [5, 6]) func loadProductsForPointOfSale_returns_false_for_hasMorePages_based_on_header(pageNumber: Int) async throws {
+        // Given a response with 5 pages
+        let remote = ProductsRemote(network: network)
+        network.responseHeaders = ["X-WP-TotalPages": "5"]
+        network.simulateResponse(requestUrlSuffix: "products", filename: "empty-data-array")
+
+        // When loading page 5 or above
+        let pagedProducts = try await remote.loadProductsForPointOfSale(
+            for: sampleSiteID,
+            pageNumber: pageNumber)
+
+        // Then there are no more pages
+        #expect(pagedProducts.hasMorePages == false)
+    }
+
+    @Test(arguments: 1...10) func loadProductsForPointOfSale_returns_hasMorePages_true_when_header_is_not_set(pageNumber: Int) async throws {
+        // Given a response with nothing in it, but no X-WP-TotalPages header
+        let remote = ProductsRemote(network: network)
+        network.responseHeaders = nil
+        network.simulateResponse(requestUrlSuffix: "products", filename: "empty-data-array")
+
+        // When loading page
+        let pagedProducts = try await remote.loadProductsForPointOfSale(
+            for: sampleSiteID,
+            pageNumber: pageNumber)
+
+        // Then we assume there are still more pages
+        #expect(pagedProducts.hasMorePages == true)
+    }
+
+    @Test func searchProductsForPointOfSale_sets_correct_parameters() async throws {
+        // Given
+        let remote = ProductsRemote(network: network)
+
+        // When
+        _ = try? await remote.searchProductsForPointOfSale(
+            for: sampleSiteID,
+            query: "search terms",
+            productTypes: [.simple, .variable],
+            pageNumber: 1)
+
+        // Then
+        let queryParametersDictionary = try #require(network.queryParametersDictionary as? [String: String])
+        #expect(queryParametersDictionary["downloadable"] == "false")
+        #expect(queryParametersDictionary["include_types"] == "simple,variable")
+        #expect(queryParametersDictionary["search"] == "search terms")
+    }
+}
