+++ Networking/NetworkingTests/Remote/ProductsRemoteTests.swift
@@ -931,122 +931,6 @@ final class ProductsRemoteTests: XCTestCase {
             $0 as? NetworkError == .notFound()
         })
     }
-
-    func test_loadProductsForPointOfSale_loads_simple_products() async throws {
-        // Given
-        let remote = ProductsRemote(network: network)
-        let expectedProductsFromResponse = 6
-
-        // When
-        network.simulateResponse(requestUrlSuffix: "products", filename: "products-load-all-type-simple")
-
-        let pagedProducts = try await remote.loadProductsForPointOfSale(for: sampleSiteID)
-
-        // Then
-        let products = pagedProducts.items
-        XCTAssertEqual(products.count, expectedProductsFromResponse)
-        for product in products {
-            XCTAssertEqual(try XCTUnwrap(product).productType, .simple)
-        }
-    }
-
-    func test_loadProductsForPointOfSale_relays_networking_error() async throws {
-        // Given
-        let remote = ProductsRemote(network: network)
-
-        // When/Then
-        await assertThrowsError({
-            let _ = try await remote.loadProductsForPointOfSale(for: sampleSiteID)
-        }, errorAssert: {
-            $0 as? NetworkError == .notFound()
-        })
-    }
-
-    func test_loadProductsForPointOfSale_when_page_has_products_then_loads_expected_products() async throws {
-        // Given
-        let remote = ProductsRemote(network: network)
-        let initialPageNumber = 1
-        let expectedProductsFromResponse = 6
-
-        // When
-        network.simulateResponse(requestUrlSuffix: "products", filename: "products-load-all-type-simple")
-
-        let pagedProducts = try await remote.loadProductsForPointOfSale(for: sampleSiteID, pageNumber: initialPageNumber)
-
-        // Then
-        let products = pagedProducts.items
-        XCTAssertEqual(products.count, expectedProductsFromResponse)
-        for product in products {
-            XCTAssertEqual(try XCTUnwrap(product).productType, .simple)
-        }
-    }
-
-    func test_loadProductsForPointOfSale_sets_correct_parameters() async throws {
-        // Given
-        let remote = ProductsRemote(network: network)
-
-        // When
-        _ = try? await remote.loadProductsForPointOfSale(for: sampleSiteID, productTypes: [.simple, .variable], pageNumber: 1)
-
-        // Then
-        let queryParametersDictionary = try XCTUnwrap(network.queryParametersDictionary)
-        XCTAssertEqual(queryParametersDictionary["downloadable"] as? String, String(false))
-        XCTAssertEqual(queryParametersDictionary["include_types"] as? String, "simple,variable")
-    }
-
-    func test_loadProductsForPointOfSale_returns_hasMorePages_based_on_header_with_case_insensitive_name() async throws {
-        // Given
-        let remote = ProductsRemote(network: network)
-        network.responseHeaders = ["X-WP-TotalPages": "5"]
-        network.simulateResponse(requestUrlSuffix: "products", filename: "empty-data-array")
-
-        // When loading page 1 to 4
-        for pageNumber in 1...4 {
-            let pagedProducts = try await remote.loadProductsForPointOfSale(for: sampleSiteID, pageNumber: pageNumber)
-
-            // Then
-            XCTAssertTrue(pagedProducts.hasMorePages)
-        }
-
-        // When loading page 17
-        let pagedProducts = try await remote.loadProductsForPointOfSale(for: sampleSiteID, pageNumber: 5)
-
-        // Then
-        XCTAssertFalse(pagedProducts.hasMorePages)
-    }
-
-    func test_loadProductsForPointOfSale_returns_hasMorePages_true_when_header_is_not_set() async throws {
-        // Given
-        let remote = ProductsRemote(network: network)
-        network.responseHeaders = nil
-        network.simulateResponse(requestUrlSuffix: "products", filename: "empty-data-array")
-
-        // When loading the first 5 pages
-        for pageNumber in 1...5 {
-            let pagedProducts = try await remote.loadProductsForPointOfSale(for: sampleSiteID, pageNumber: pageNumber)
-
-            // Then
-            XCTAssertTrue(pagedProducts.hasMorePages)
-        }
-    }
-
-    func test_searchProductsForPointOfSale_sets_correct_parameters() async throws {
-        // Given
-        let remote = ProductsRemote(network: network)
-
-        // When
-        _ = try? await remote.searchProductsForPointOfSale(
-            for: sampleSiteID,
-            query: "search terms",
-            productTypes: [.simple, .variable],
-            pageNumber: 1)
-
-        // Then
-        let queryParametersDictionary = try XCTUnwrap(network.queryParametersDictionary)
-        XCTAssertEqual(queryParametersDictionary["downloadable"] as? String, String(false))
-        XCTAssertEqual(queryParametersDictionary["include_types"] as? String, "simple,variable")
-        XCTAssertEqual(queryParametersDictionary["search"] as? String, "search terms")
-    }
 }
 
 // MARK: - Private Helpers
